+++ Samples/MatrixKitSample/MXKSampleJSQMessagesViewController.m
@@ -413,7 +413,7 @@
     }
     
     // Compute the member avatar URL
-    MXRoomMember *roomMember = [roomDataSource.room.state memberWithUserId:messageData.senderId];
+    MXRoomMember *roomMember = [roomDataSource.roomState.members memberWithUserId:messageData.senderId];
     
     NSString *avatarUrl = [roomDataSource.mxSession.matrixRestClient urlOfContentThumbnail:roomMember.avatarUrl toFitViewSize:cell.avatarImageView.frame.size withMethod:MXThumbnailingMethodCrop];
     if (!avatarUrl)