+++ CHANGES.rst
@@ -11,6 +11,8 @@ Bug fix:
 API break:
  * MXKContactManager: Remove the privateMatrixContacts method.
  * MXKSearchCellDataStoring: Replace initWithSearchResult by async cellDataWithSearchResult.
+ * MXKRoomDataSourceManager: The roomDataSourceForRoom method is now asynchronous.
+ * MXKRoomDataSourceManager: closeRoomDataSource has been replaced by closeRoomDataSourceWithRoomId.
 
 Changes in MatrixKit in 0.7.15 (2018-07-03)
 ==========================================